Revert "implement a way to group the exercises by their bodypart"
This reverts commit fec6418192103815998ee7a5c26b689773e8b0f6.

+++ GymLogger/FetchControllerBase.m
@@ -110,7 +110,7 @@
 }
 
 - (NSString *)tableView:(UITableView *)tableView titleForHeaderInSection:(NSInteger)section {
-    return [[self.fetchedResultsController sections][(NSUInteger) section] name];
+    return [[self.fetchedResultsController sections][section] name];
 }
 
 @end
